Make Local Archive Name Dynamic
Prior to this commit the local_archive name was static.  This could
cause corrective changes for something that was intentional because the
source file was updated but the local target was not.  This commit
determines the local_archive name by using basename() against the
dist_file param.  Also updated tests to reflect this change.